requirements.txt
alembic==1.1.0
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2==2.8.3
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-telegram-bot==12.0.0
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
tornado==6.0.3
urllib3==1.25.3
uWSGI==2.0.18
Werkzeug==0.15.6
yandex-money-sdk==0.1.3
